move inventory functions to separate file
