+++ ORV Text Adventure/Test.py
@@ -10,16 +10,16 @@ something = item.Item("something", False, item.Grade.A, basicAbility)
 somethingElse = item.Item("somethingElse", True, item.Grade.A, basicAbility)
 aThing = item.Item("aThing", True, item.Grade.A, basicAbility)
 
-player.displayInventory()
-player.addToInventory(something, 1)
-player.addToInventory(somethingElse, 9)
-player.displayInventory()
-player.removeFromInventory(somethingElse, 1)
-player.removeFromInventory(something, 2)
-player.displayInventory()
-player.removeFromInventory(somethingElse, -1)
-player.removeFromInventory(aThing, 1)
-player.displayInventory()
-player.accessInventory()
+player.inventory.displayInventory()
+player.inventory.addToInventory(something, 1)
+player.inventory.addToInventory(somethingElse, 9)
+player.inventory.displayInventory()
+player.inventory.removeFromInventory(somethingElse, 1)
+player.inventory.removeFromInventory(something, 2)
+player.inventory.displayInventory()
+player.inventory.removeFromInventory(somethingElse, -1)
+player.inventory.removeFromInventory(aThing, 1)
+player.inventory.displayInventory()
+player.inventory.accessInventory()
 # Equipped item tests
 # somethingElse.activateAbility()
