+++ ORV Text Adventure/Text.py
@@ -1,0 +1,2 @@
+def displayLine():
+        print("__________________________________________________________________________________")
